requirements.txt
setuptools
wheel
build
absl-py==2.1.0
accelerate==1.1.1
aiohappyeyeballs==2.4.3
aiohttp==3.11.7
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.11.0
appdirs==1.4.4
appnope==0.1.4
asttokens==2.4.1
attrs==23.1.0
backoff==2.2.1
bcrypt==5.0.0
beautifulsoup4==4.12.2
blinker==1.9.0
build==1.3.0
cachetools==6.2.1
certifi==2023.11.17
cffi==2.0.0
cfgv==3.4.0
charset-normalizer==3.3.2
chromadb==1.1.1
click==8.2.1
cloudpickle==3.0.0
coloredlogs==15.0.1
comm==0.2.2
contourpy==1.3.0
crewai==1.2.1
cryptography==46.0.3
cycler==0.12.1
dash==3.2.0
dash-bootstrap-components==2.0.4
dash-iconify==0.1.2
dataclasses-json==0.6.7
datasets==3.1.0
debugpy==1.8.5
decorator==5.1.1
dill==0.3.8
diskcache==5.6.3
distlib==0.4.0
distro==1.9.0
docstring_parser==0.17.0
dotenv==0.9.9
duckdb==1.4.1
durationpy==0.10
et_xmlfile==2.0.0
evaluate==0.4.3
exceptiongroup==1.2.0
executing==2.1.0
faiss-cpu==1.12.0
Farama-Notifications==0.0.4
fastuuid==0.14.0
filelock==3.16.1
Flask==3.1.2
flatbuffers==25.9.23
fonttools==4.54.1
frozenlist==1.5.0
fsspec==2024.9.0
google-api-core==2.28.1
google-api-python-client==2.187.0
google-auth==2.41.1
google-auth-httplib2==0.2.1
google-auth-oauthlib==1.2.3
googleapis-common-protos==1.71.0
grpcio==1.70.0
gunicorn==23.0.0
gymnasium==0.29.1
h11==0.16.0
httpcore==1.0.9
httplib2==0.31.0
httptools==0.7.1
httpx==0.28.1
httpx-sse==0.4.3
huggingface-hub==0.26.2
humanfriendly==10.0
identify==2.6.15
idna==3.6
importlib_metadata==8.7.0
importlib_resources==6.5.2
instructor==1.12.0
ipykernel==6.29.5
ipython==8.27.0
itsdangerous==2.2.0
jedi==0.19.1
Jinja2==3.1.4
jiter==0.10.0
joblib==1.5.2
json5==0.12.1
json_repair==0.25.2
jsonpatch==1.33
jsonpickle==4.1.1
jsonpointer==3.0.0
jsonref==1.1.0
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
jupyter_client==8.6.2
jupyter_core==5.7.2
kiwisolver==1.4.7
kubernetes==34.1.0
litellm==1.79.0
lxml==4.9.3
Markdown==3.7
markdown-it-py==4.0.0
markdown2==2.5.4
MarkupSafe==2.1.5
marshmallow==3.26.1
matplotlib==3.9.2
matplotlib-inline==0.1.7
mcp==1.20.0
mdurl==0.1.2
mmh3==5.2.0
mpmath==1.3.0
multidict==6.1.0
multiprocess==0.70.16
mypy_extensions==1.1.0
narwhals==2.1.2
nest-asyncio==1.6.0
networkx==3.3
nodeenv==1.9.1
numpy==1.26.2
oauthlib==3.3.1
onnxruntime==1.23.2
openai==1.109.1
opencv-python==4.11.0.86
openpyxl==3.1.5
opentelemetry-api==1.38.0
opentelemetry-exporter-otlp-proto-common==1.38.0
opentelemetry-exporter-otlp-proto-grpc==1.38.0
opentelemetry-exporter-otlp-proto-http==1.38.0
opentelemetry-proto==1.38.0
opentelemetry-sdk==1.38.0
opentelemetry-semantic-conventions==0.59b0
orjson==3.11.3
ormsgpack==1.11.0
outcome==1.3.0.post0
overrides==7.7.0
packaging==24.1
pandas==2.1.3
parso==0.8.4
pdfminer.six==20250506
pdfplumber==0.11.7
pexpect==4.9.0
pillow==10.4.0
platformdirs==4.2.2
plotly==6.3.0
portalocker==2.7.0
posthog==5.4.0
pre_commit==4.3.0
prompt_toolkit==3.0.47
propcache==0.2.0
proto-plus==1.26.1
protobuf==5.29.3
psutil==6.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
pyarrow==18.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pybase64==1.4.2
pycparser==2.23
pydantic==2.12.1
pydantic-settings==2.11.0
pydantic_core==2.41.3
Pygments==2.18.0
PyJWT==2.10.1
pyparsing==3.1.4
pypdf==6.1.1
PyPDF2==3.0.1
pypdfium2==5.0.0
PyPika==0.48.9
pyproject_hooks==1.2.0
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==1.2.1
python-multipart==0.0.20
pytz==2023.3.post1
pyvis==0.3.2
PyYAML==6.0.2
pyzmq==26.2.0
referencing==0.37.0
regex==2024.11.6
requests==2.32.5
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
retrying==1.4.2
rich==14.2.0
rpds-py==0.28.0
rsa==4.9.1
safetensors==0.4.5
scikit-learn==1.7.2
scipy==1.16.2
selenium==4.15.2
selenium-stealth==1.0.6
sentence-transformers==5.1.1
setuptools==75.1.0
shellingham==1.5.4
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.5
SQLAlchemy==2.0.44
sse-starlette==3.0.3
stack-data==0.6.3
starlette==0.49.1
sympy==1.13.1
tabulate==0.9.0
tenacity==8.5.0
tensorboard==2.18.0
tensorboard-data-server==0.7.2
termcolor==2.4.0
threadpoolctl==3.6.0
tiktoken==0.12.0
tokenizers==0.20.3
tomli==2.3.0
tomli_w==1.2.0
torch==2.5.1
torch-tb-profiler==0.4.3
torchaudio==2.5.1
torchvision==0.20.1
tornado==6.4.1
tqdm==4.66.5
traitlets==5.14.3
transformers==4.46.3
trio==0.23.1
trio-websocket==0.11.1
typer==0.20.0
typing-inspect==0.9.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2023.3
undetected-chromedriver==3.5.4
uritemplate==4.2.0
urllib3==2.1.0
uv==0.9.7
uvicorn==0.38.0
uvloop==0.22.1
virtualenv==20.35.4
watchfiles==1.1.1
wcwidth==0.2.13
websocket-client==1.9.0
websockets==12.0
Werkzeug==3.1.3
wsproto==1.2.0
xxhash==3.5.0
yarl==1.18.0
zipp==3.23.0
zstandard==0.25.0
